Extract dialect-related utilities from test_utils.rs and re-export
